feat(pm): download in-production size-PIC report from the style page
Adds a "Download in-production report" button on /pm/style/:style that exports
the size-wise PIC report (one row per lot+size, all stage columns) for every lot
currently in production across all styles — same format as the operator
dashboard's /dashboard/pic-size-report, output as PICReport-InProduction-BySize.xlsx.

To keep the two reports from drifting, the PIC helpers are extracted verbatim from
routes/operatorRoutes.js into a shared utils/picSizeReport.js: isDenimLot,
parseLotRemark, fmtIST, daysSince, classifyStage, buildEnhancedRow,
PIC_REPORT_V2_COLUMNS, fetchLotEventAggregates, fetchLotSizeEventSums,
getDepartmentStatuses, filterByDept — plus new buildPicSizeRows()/buildPicSizeWorkbook().
operatorRoutes.js now imports these (behaviour unchanged).

Notes:
- daysSince(): operatorRoutes had two top-level declarations; in sloppy mode the
  LATER one won, so buildEnhancedRow effectively used it. The module carries that
  exact version so output is byte-identical; both dead copies removed from the route.
- "In production" = lot cut within the last 120 days with net (cut - dispatched) > 0,
  matching the in-flight window utils/onOrder.js uses for the PM dashboard.
- The all-in-production path raises the query cap 5000 -> 50000 so it doesn't silently
  drop older in-production lots (120d window is ~6.5k lot x size rows and growing).

Verified against prod (read-only): 6,395 in-production rows across 515 SKUs, the
net-undispatched invariant holds for every row, and the workbook builds with the
identical 65-column layout. Both route modules load and register their routers.
