+++ routes/operatorRoutes.js
@@ -967,948 +967,19 @@ router.get("/security-logs", isAuthenticated, isOperator, async (req, res) => {
  * 6) PIC Report – corrected chain
  **************************************************/
 // Quick helper: isDenimLot - checks lot type using flow_type, is_denim_cutter, or lot prefix
-function isDenimLot(lotOrLotNo, isDenimCutter = null, flowType = null) {
-  // If called with lot object that has flow_type or is_denim_cutter
-  if (typeof lotOrLotNo === 'object' && lotOrLotNo !== null) {
-    if (lotOrLotNo.flow_type) return lotOrLotNo.flow_type === 'denim';
-    if (lotOrLotNo.is_denim_cutter !== null && lotOrLotNo.is_denim_cutter !== undefined) {
-      return lotOrLotNo.is_denim_cutter === 1;
-    }
-    // Fallback to lot prefix
-    const up = (lotOrLotNo.lot_no || '').toUpperCase();
-    return up.startsWith("AK") || up.startsWith("UM");
-  }
-  // If called with flowType or isDenimCutter params
-  if (flowType) return flowType === 'denim';
-  if (isDenimCutter !== null && isDenimCutter !== undefined) return isDenimCutter === 1;
-  // Fallback: called with just lot_no string
-  const up = String(lotOrLotNo).toUpperCase();
-  return up.startsWith("AK") || up.startsWith("UM");
-}
+/**************************************************
+ * 6) PIC Report helpers — extracted to utils/picSizeReport.js
+ **************************************************/
+const {
+  isDenimLot,
+  buildEnhancedRow,
+  PIC_REPORT_V2_COLUMNS,
+  fetchLotEventAggregates,
+  fetchLotSizeEventSums,
+  getDepartmentStatuses,
+  filterByDept,
+} = require("../utils/picSizeReport");
 
-// ---------- PIC-Report v2 helpers (user-friendly schema) ----------
-function parseLotRemark(remark) {
-  if (!remark) return { externalLotNo: '', sortNo: '' };
-  const s = String(remark);
-  const lotMatch  = s.match(/LOT\s*N(?:O|UMBER)?\s*[:#-]?\s*([A-Za-z0-9_\-\/]+)/i);
-  const sortMatch = s.match(/SORT\s*N(?:O|UMBER)?\s*[:#-]?\s*([A-Za-z0-9_\-\/]+)/i);
-  return {
-    externalLotNo: lotMatch  ? lotMatch[1]  : '',
-    sortNo:        sortMatch ? sortMatch[1] : ''
-  };
-}
-
-function fmtIST(d) {
-  if (!d) return '';
-  const dt = new Date(d);
-  if (isNaN(dt.getTime())) return '';
-  return dt.toLocaleString('en-GB', { timeZone: 'Asia/Kolkata' }).replace(/\//g, '-');
-}
-
-function daysSince(d, ref = new Date()) {
-  if (!d) return '';
-  const dt = new Date(d);
-  if (isNaN(dt.getTime())) return '';
-  const ms = ref.getTime() - dt.getTime();
-  return Math.max(0, Math.floor(ms / 86400000));
-}
-
-// Per-stage classifier. Returns a status that describes ONLY this stage.
-//   isApplicable=false      => N/A (e.g. assembly/washing for hosiery)
-//   no assignment           => Not Started
-//   is_approved is null     => Pending Approval
-//   is_approved == 0        => Denied
-//   approved & out=0        => In Progress
-//   approved & out >= in    => Completed
-//   approved & 0<out<in     => Partial
-function classifyStage({ assign, inQty, outQty, nextAssign, isApplicable }) {
-  if (!isApplicable) {
-    return { status: 'N/A', pending: 0, inline: '' };
-  }
-  if (!assign) {
-    return { status: 'Not Started', pending: Math.max(0, inQty || 0), inline: '' };
-  }
-  const approved = (assign.is_approved !== undefined) ? assign.is_approved : assign.isApproved;
-  if (approved === null || approved === undefined) {
-    return { status: 'Pending Approval', pending: Math.max(0, (inQty || 0) - (outQty || 0)), inline: '' };
-  }
-  if (Number(approved) === 0) {
-    return { status: 'Denied', pending: Math.max(0, (inQty || 0) - (outQty || 0)), inline: '' };
-  }
-  // approved
-  if (!outQty || outQty <= 0) {
-    return { status: 'In Progress', pending: Math.max(0, inQty || 0), inline: nextAssign ? 'Yes' : '' };
-  }
-  if (inQty > 0 && outQty >= inQty) {
-    return { status: 'Completed', pending: 0, inline: nextAssign ? '' : 'Yes' };
-  }
-  return { status: 'Partial', pending: Math.max(0, inQty - outQty), inline: nextAssign ? 'Yes' : '' };
-}
-
-// Build a flat, user-friendly row: per-stage In/Out/Pending/Status/Inline + a top-level rollup.
-function buildEnhancedRow({
-  lot, isDenim, totalCut, sums, assigns,
-  rewash = { requested: 0, pending: 0, completed: 0 },
-  rejects = {}
-}) {
-  // Back-compat shim: callers may still pass `rewashQty` (a number for pending).
-  if (typeof arguments[0].rewashQty === 'number') {
-    rewash = { requested: 0, pending: arguments[0].rewashQty, completed: 0 };
-  }
-  const { stitchedQty, assembledQty, washedQty, washingInQty, finishedQty } = sums;
-  const { stAssign, asmAssign, washAssign, washInAssign, finAssign } = assigns;
-
-  // Stage flows:
-  //   DENIM:    Cut → Stitch → Assembly → Washing → WashIn → Finishing
-  //   HOSIERY:  Cut → Stitch → Finishing
-  const stitch = classifyStage({
-    assign: stAssign,
-    inQty:  totalCut,
-    outQty: stitchedQty,
-    nextAssign: isDenim ? asmAssign : finAssign,
-    isApplicable: true
-  });
-  const assembly = classifyStage({
-    assign: asmAssign,
-    inQty:  stitchedQty,
-    outQty: assembledQty,
-    nextAssign: washAssign,
-    isApplicable: isDenim
-  });
-  const washing = classifyStage({
-    assign: washAssign,
-    inQty:  assembledQty,
-    outQty: washedQty,
-    nextAssign: washInAssign,
-    isApplicable: isDenim
-  });
-  const washIn = classifyStage({
-    assign: washInAssign,
-    inQty:  washedQty,
-    outQty: washingInQty,
-    nextAssign: finAssign,
-    isApplicable: isDenim
-  });
-  const finishing = classifyStage({
-    assign: finAssign,
-    inQty:  isDenim ? washingInQty : stitchedQty,
-    outQty: finishedQty,
-    nextAssign: null,
-    isApplicable: true
-  });
-
-  // Determine current stage: first non-Completed, non-NA in the chain.
-  const chain = isDenim
-    ? [['Stitching', stitch, stAssign],
-       ['Assembly',  assembly, asmAssign],
-       ['Washing',   washing,  washAssign],
-       ['Wash-In',   washIn,   washInAssign],
-       ['Finishing', finishing, finAssign]]
-    : [['Stitching', stitch, stAssign],
-       ['Finishing', finishing, finAssign]];
-
-  let currentStage = 'Done';
-  let currentPending = 0;
-  let stageStartTs = lot.created_at || null;
-  for (const [name, info, assign] of chain) {
-    if (info.status === 'N/A') continue;
-    if (info.status === 'Completed') {
-      // advance start ts to this stage's approval
-      if (assign && assign.approved_on) stageStartTs = assign.approved_on;
-      continue;
-    }
-    currentStage = name;
-    currentPending = info.pending;
-    if (assign && assign.assigned_on) stageStartTs = assign.assigned_on;
-    break;
-  }
-
-  const { externalLotNo: parsedExternalLotNo, sortNo } = parseLotRemark(lot.remark);
-  // Prefer the authoritative, backfilled manual_lot_number column; fall back to
-  // the value parsed live from the remark for any lot not yet backfilled.
-  const externalLotNo = (lot.manual_lot_number && String(lot.manual_lot_number).trim())
-    ? String(lot.manual_lot_number).trim()
-    : parsedExternalLotNo;
-  const opName = a => (a && a.opName) ? a.opName : '';
-
-  return {
-    // identification
-    lotNo: lot.lot_no,
-    externalLotNo,
-    sortNo,
-    fabricType: lot.fabric_type || '',
-    sku: lot.sku,
-    lotType: isDenim ? 'Denim' : 'Hosiery',
-    createdAt: lot.created_at
-      ? new Date(lot.created_at).toLocaleDateString('en-GB', { timeZone: 'Asia/Kolkata' }).replace(/\//g, '-')
-      : '',
-    manualCuttingDate: lot.manual_cutting_date
-      ? new Date(lot.manual_cutting_date).toLocaleDateString('en-GB', { timeZone: 'Asia/Kolkata' }).replace(/\//g, '-')
-      : '',
-    daysSinceCreated: daysSince(lot.created_at),
-    totalCut,
-    remark: lot.remark || '',
-
-    // rollup
-    currentStage,
-    currentPendingQty: currentPending,
-    daysInStage: daysSince(stageStartTs),
-
-    // stitching
-    stitchOp:        opName(stAssign),
-    stitchAssignedOn: fmtIST(stAssign && stAssign.assigned_on),
-    stitchApprovedOn: fmtIST(stAssign && stAssign.approved_on),
-    stitchInQty:     totalCut,
-    stitchOutQty:    stitchedQty,
-    stitchPendingQty: stitch.pending,
-    stitchStatus:    stitch.status,
-    stitchInline:    stitch.inline,
-
-    // assembly
-    assemblyOp:        isDenim ? opName(asmAssign) : '—',
-    assemblyAssignedOn: isDenim ? fmtIST(asmAssign && asmAssign.assigned_on) : '—',
-    assemblyApprovedOn: isDenim ? fmtIST(asmAssign && asmAssign.approved_on) : '—',
-    assemblyInQty:     isDenim ? stitchedQty  : '—',
-    assemblyOutQty:    isDenim ? assembledQty : '—',
-    assemblyPendingQty: isDenim ? assembly.pending : '—',
-    assemblyStatus:    assembly.status,
-    assemblyInline:    isDenim ? assembly.inline : '—',
-
-    // washing
-    washingOp:        isDenim ? opName(washAssign) : '—',
-    washingAssignedOn: isDenim ? fmtIST(washAssign && washAssign.assigned_on) : '—',
-    washingApprovedOn: isDenim ? fmtIST(washAssign && washAssign.approved_on) : '—',
-    washingInQty_in:  isDenim ? assembledQty : '—',
-    washingOutQty:    isDenim ? washedQty    : '—',
-    washingPendingQty: isDenim ? washing.pending : '—',
-    washingStatus:    washing.status,
-    washingInline:    isDenim ? washing.inline : '—',
-
-    // wash-in
-    washInOp:        isDenim ? opName(washInAssign) : '—',
-    washInAssignedOn: isDenim ? fmtIST(washInAssign && washInAssign.assigned_on) : '—',
-    washInApprovedOn: isDenim ? fmtIST(washInAssign && washInAssign.approved_on) : '—',
-    washInInQty:     isDenim ? washedQty    : '—',
-    washInOutQty:    isDenim ? washingInQty : '—',
-    washInPendingQty: isDenim ? washIn.pending : '—',
-    washInStatus:    washIn.status,
-    washInInline:    isDenim ? washIn.inline : '—',
-
-    // rewash (pieces sent back from wash-in to washing for re-processing)
-    rewashRequestedQty: rewash.requested || 0,
-    rewashPendingQty:   rewash.pending   || 0,
-    rewashCompletedQty: rewash.completed || 0,
-
-    // rejects per stage (pieces removed from production due to defects)
-    stitchRejectQty:     rejects.stitching   ? rejects.stitching.pieces   : 0,
-    stitchRejectReasons: rejects.stitching   ? rejects.stitching.reasons  : '',
-    washInRejectQty:     rejects.washing_in  ? rejects.washing_in.pieces  : 0,
-    washInRejectReasons: rejects.washing_in  ? rejects.washing_in.reasons : '',
-    finishingRejectQty:     rejects.finishing ? rejects.finishing.pieces  : 0,
-    finishingRejectReasons: rejects.finishing ? rejects.finishing.reasons : '',
-    totalRejectQty:
-      ((rejects.stitching  && rejects.stitching.pieces)  || 0) +
-      ((rejects.washing_in && rejects.washing_in.pieces) || 0) +
-      ((rejects.finishing  && rejects.finishing.pieces)  || 0),
-
-    // finishing
-    finishingOp:        opName(finAssign),
-    finishingAssignedOn: fmtIST(finAssign && finAssign.assigned_on),
-    finishingApprovedOn: fmtIST(finAssign && finAssign.approved_on),
-    finishingInQty:     isDenim ? washingInQty : stitchedQty,
-    finishingOutQty:    finishedQty,
-    finishingPendingQty: finishing.pending,
-    finishingStatus:    finishing.status
-  };
-}
-
-const PIC_REPORT_V2_COLUMNS = [
-  { header: 'Lot No',              key: 'lotNo',             width: 14 },
-  { header: 'Manual Lot No',       key: 'externalLotNo',     width: 14 },
-  { header: 'Fabric Type',         key: 'fabricType',        width: 14 },
-  { header: 'SKU',                 key: 'sku',               width: 22 },
-  { header: 'Lot Type',            key: 'lotType',           width: 9  },
-  { header: 'Created At',          key: 'createdAt',         width: 12 },
-  { header: 'Manual Cutting Date', key: 'manualCuttingDate', width: 14 },
-  { header: 'Days Since Created',  key: 'daysSinceCreated',  width: 10 },
-  { header: 'Total Cut',           key: 'totalCut',          width: 10 },
-  { header: 'Current Stage',       key: 'currentStage',      width: 14 },
-  { header: 'Current Pending Qty', key: 'currentPendingQty', width: 12 },
-  { header: 'Days In Stage',       key: 'daysInStage',       width: 10 },
-  { header: 'Remark',              key: 'remark',            width: 26 },
-
-  { header: 'Stitch Operator',     key: 'stitchOp',          width: 14 },
-  { header: 'Stitch Assigned On',  key: 'stitchAssignedOn',  width: 19 },
-  { header: 'Stitch Approved On',  key: 'stitchApprovedOn',  width: 19 },
-  { header: 'Stitch In Qty',       key: 'stitchInQty',       width: 11 },
-  { header: 'Stitch Out Qty',      key: 'stitchOutQty',      width: 11 },
-  { header: 'Stitch Pending Qty',  key: 'stitchPendingQty',  width: 12 },
-  { header: 'Stitch Status',       key: 'stitchStatus',      width: 16 },
-  { header: 'Stitch Inline?',      key: 'stitchInline',      width: 9  },
-
-  { header: 'Assembly Operator',     key: 'assemblyOp',          width: 14 },
-  { header: 'Assembly Assigned On',  key: 'assemblyAssignedOn',  width: 19 },
-  { header: 'Assembly Approved On',  key: 'assemblyApprovedOn',  width: 19 },
-  { header: 'Assembly In Qty',       key: 'assemblyInQty',       width: 11 },
-  { header: 'Assembly Out Qty',      key: 'assemblyOutQty',      width: 11 },
-  { header: 'Assembly Pending Qty',  key: 'assemblyPendingQty',  width: 12 },
-  { header: 'Assembly Status',       key: 'assemblyStatus',      width: 16 },
-  { header: 'Assembly Inline?',      key: 'assemblyInline',      width: 9  },
-
-  { header: 'Washing Operator',     key: 'washingOp',          width: 14 },
-  { header: 'Washing Assigned On',  key: 'washingAssignedOn',  width: 19 },
-  { header: 'Washing Approved On',  key: 'washingApprovedOn',  width: 19 },
-  { header: 'Washing In Qty',       key: 'washingInQty_in',    width: 11 },
-  { header: 'Washing Out Qty',      key: 'washingOutQty',      width: 11 },
-  { header: 'Washing Pending Qty',  key: 'washingPendingQty',  width: 12 },
-  { header: 'Washing Status',       key: 'washingStatus',      width: 16 },
-  { header: 'Washing Inline?',      key: 'washingInline',      width: 9  },
-
-  { header: 'Wash-In Operator',     key: 'washInOp',           width: 14 },
-  { header: 'Wash-In Assigned On',  key: 'washInAssignedOn',   width: 19 },
-  { header: 'Wash-In Approved On',  key: 'washInApprovedOn',   width: 19 },
-  { header: 'Wash-In In Qty',       key: 'washInInQty',        width: 11 },
-  { header: 'Wash-In Out Qty',      key: 'washInOutQty',       width: 11 },
-  { header: 'Wash-In Pending Qty',  key: 'washInPendingQty',   width: 12 },
-  { header: 'Wash-In Status',       key: 'washInStatus',       width: 16 },
-  { header: 'Wash-In Inline?',      key: 'washInInline',       width: 9  },
-
-  { header: 'Rewash Requested',     key: 'rewashRequestedQty', width: 11 },
-  { header: 'Rewash Pending',       key: 'rewashPendingQty',   width: 11 },
-  { header: 'Rewash Completed',     key: 'rewashCompletedQty', width: 11 },
-
-  { header: 'Finishing Operator',     key: 'finishingOp',          width: 14 },
-  { header: 'Finishing Assigned On',  key: 'finishingAssignedOn',  width: 19 },
-  { header: 'Finishing Approved On',  key: 'finishingApprovedOn',  width: 19 },
-  { header: 'Finishing In Qty',       key: 'finishingInQty',       width: 11 },
-  { header: 'Finishing Out Qty',      key: 'finishingOutQty',      width: 11 },
-  { header: 'Finishing Pending Qty',  key: 'finishingPendingQty',  width: 12 },
-  { header: 'Finishing Status',       key: 'finishingStatus',      width: 16 },
-
-  { header: 'Stitch Rejects',         key: 'stitchRejectQty',         width: 11 },
-  { header: 'Stitch Reject Reasons',  key: 'stitchRejectReasons',     width: 28 },
-  { header: 'Wash-In Rejects',        key: 'washInRejectQty',         width: 11 },
-  { header: 'Wash-In Reject Reasons', key: 'washInRejectReasons',     width: 28 },
-  { header: 'Finishing Rejects',      key: 'finishingRejectQty',      width: 11 },
-  { header: 'Finishing Reject Reasons', key: 'finishingRejectReasons', width: 28 },
-  { header: 'Total Rejects',          key: 'totalRejectQty',          width: 11 }
-];
-
-// ---------------------------------------------------------------------------
-// Event-sourced aggregates (truth source for PIC reports).
-//
-// The legacy *_assignments / *_data tables are only partially kept in sync —
-// roughly half of recent activity flows through the *_events tables instead.
-// These helpers read directly from *_events / *_event_sizes so the reports
-// reflect real lot location (e.g. AK5237 sitting inline at washer ADS even
-// though washing_assignments has no row for it).
-//
-// Mapping back to the legacy shape:
-//   stitchedQty   = SUM(stitching_events.complete)
-//   assembledQty  = SUM(jeans_assembly_events.complete)
-//   washedQty     = SUM(washing_events.complete)
-//   washingInQty  = SUM(washing_in_events.complete)
-//   finishedQty   = SUM(finishing_events.complete)
-//
-//   stitchMap[lot] / asmMap / washMap / winMap / finMap = a synthetic
-//   "assignment-like" object built from the latest approve event for the
-//   stage (is_approved=1, assigned_on=approved_on=latest approve ts,
-//   opName=latest approver username). null when no approve event yet.
-// ---------------------------------------------------------------------------
-async function fetchLotEventAggregates(lotNos = []) {
-  if (!lotNos.length) {
-    return {
-      lotSumsMap: {}, stitchMap: {}, asmMap: {}, washMap: {}, winMap: {}, finMap: {}
-    };
-  }
-  const sorted = lotNos.slice().sort();
-  const cacheKey = `lotAggEv-${sorted.length}-${sorted[0]}-${sorted[sorted.length - 1]}`;
-  return cache.fetchCached(cacheKey, async () => {
-    const aggSql = (table, key) => `
-      SELECT '${key}' AS stage, cl.lot_no,
-             SUM(CASE WHEN e.event_type='complete' THEN e.pieces ELSE 0 END) AS completed
-        FROM ${table} e
-        JOIN cutting_lots cl ON cl.id = e.cutting_lot_id
-       WHERE cl.lot_no IN (?)
-       GROUP BY cl.lot_no
-    `;
-    const aggQ = pool.query(
-      [
-        aggSql('stitching_events',       'stitched'),
-        aggSql('jeans_assembly_events',  'assembled'),
-        aggSql('washing_events',         'washed'),
-        aggSql('washing_in_events',      'washing_in'),
-        aggSql('finishing_events',       'finished'),
-      ].join('\nUNION ALL\n'),
-      [lotNos, lotNos, lotNos, lotNos, lotNos]
-    );
-
-    // Per-stage approve events, ordered so the first row per lot is the latest.
-    const opSql = (table) => `
-      SELECT cl.lot_no, e.created_at, u.username AS opName, e.operator_id
-        FROM ${table} e
-        JOIN cutting_lots cl ON cl.id = e.cutting_lot_id
-        JOIN users u ON u.id = e.operator_id
-       WHERE cl.lot_no IN (?) AND e.event_type = 'approve'
-       ORDER BY e.created_at DESC, e.id DESC
-    `;
-    const stApprovesQ  = pool.query(opSql('stitching_events'),       [lotNos]);
-    const asmApprovesQ = pool.query(opSql('jeans_assembly_events'),  [lotNos]);
-    const washApprovesQ= pool.query(opSql('washing_events'),         [lotNos]);
-    const winApprovesQ = pool.query(opSql('washing_in_events'),      [lotNos]);
-    const finApprovesQ = pool.query(opSql('finishing_events'),       [lotNos]);
-
-    const [
-      [aggRows],
-      [stApproves],
-      [asmApproves],
-      [washApproves],
-      [winApproves],
-      [finApproves],
-    ] = await Promise.all([aggQ, stApprovesQ, asmApprovesQ, washApprovesQ, winApprovesQ, finApprovesQ]);
-
-    const lotSumsMap = {};
-    lotNos.forEach(ln => {
-      lotSumsMap[ln] = { stitchedQty:0, assembledQty:0, washedQty:0, washingInQty:0, finishedQty:0 };
-    });
-    for (const r of aggRows) {
-      const m = lotSumsMap[r.lot_no];
-      if (!m) continue;
-      const completed = parseFloat(r.completed) || 0;
-      switch (r.stage) {
-        case 'stitched':   m.stitchedQty   = completed; break;
-        case 'assembled':  m.assembledQty  = completed; break;
-        case 'washed':     m.washedQty     = completed; break;
-        case 'washing_in': m.washingInQty  = completed; break;
-        case 'finished':   m.finishedQty   = completed; break;
-      }
-    }
-
-    const latestApproveMap = (rows) => {
-      const m = {};
-      for (const r of rows) {
-        if (m[r.lot_no]) continue; // first occurrence is latest (DESC order)
-        m[r.lot_no] = {
-          is_approved: 1,
-          assigned_on: r.created_at,
-          approved_on: r.created_at,
-          opName:      r.opName,
-          user_id:     r.operator_id,
-        };
-      }
-      return m;
-    };
-
-    return {
-      lotSumsMap,
-      stitchMap: latestApproveMap(stApproves),
-      asmMap:    latestApproveMap(asmApproves),
-      washMap:   latestApproveMap(washApproves),
-      winMap:    latestApproveMap(winApproves),
-      finMap:    latestApproveMap(finApproves),
-    };
-  });
-}
-
-// Per-(lot, size) completed pieces by stage, from *_event_sizes.
-async function fetchLotSizeEventSums(lotNos = []) {
-  if (!lotNos.length) return {};
-  const sorted = lotNos.slice().sort();
-  const cacheKey = `lotSizeEv-${sorted.length}-${sorted[0]}-${sorted[sorted.length - 1]}`;
-  return cache.fetchCached(cacheKey, async () => {
-    const sizeSql = (eventsTbl, sizesTbl, key) => `
-      SELECT '${key}' AS stage, cl.lot_no, s.size_label,
-             COALESCE(SUM(s.pieces),0) AS completed
-        FROM ${sizesTbl} s
-        JOIN ${eventsTbl} e ON e.id = s.event_id
-        JOIN cutting_lots cl ON cl.id = e.cutting_lot_id
-       WHERE cl.lot_no IN (?) AND e.event_type='complete'
-       GROUP BY cl.lot_no, s.size_label
-    `;
-    const [rows] = await pool.query(
-      [
-        sizeSql('stitching_events',      'stitching_event_sizes',      'stitched'),
-        sizeSql('jeans_assembly_events', 'jeans_assembly_event_sizes', 'assembled'),
-        sizeSql('washing_events',        'washing_event_sizes',        'washed'),
-        sizeSql('washing_in_events',     'washing_in_event_sizes',     'washing_in'),
-        sizeSql('finishing_events',      'finishing_event_sizes',      'finished'),
-      ].join('\nUNION ALL\n'),
-      [lotNos, lotNos, lotNos, lotNos, lotNos]
-    );
-    const map = {}; // key = `${lot_no}|${size_label}`
-    for (const r of rows) {
-      const k = `${r.lot_no}|${r.size_label}`;
-      if (!map[k]) map[k] = { stitchedQty:0, assembledQty:0, washedQty:0, washingInQty:0, finishedQty:0 };
-      const completed = parseFloat(r.completed) || 0;
-      switch (r.stage) {
-        case 'stitched':   map[k].stitchedQty   = completed; break;
-        case 'assembled':  map[k].assembledQty  = completed; break;
-        case 'washed':     map[k].washedQty     = completed; break;
-        case 'washing_in': map[k].washingInQty  = completed; break;
-        case 'finished':   map[k].finishedQty   = completed; break;
-      }
-    }
-    return map;
-  });
-}
-
-/**
- * The chain logic you want:
- * DENIM: Cut → Stitching → Assembly → Washing → WashingIn → Finishing
- * NON-DENIM: Cut → Stitching → Finishing
- * 
- * If a step is not assigned or partial, we say all subsequent steps = "In <that step>".
- */
-function getDepartmentStatuses({
-  isDenim,
-  totalCut,
-  stitchedQty,
-  assembledQty,
-  washedQty,
-  washingInQty,
-  finishedQty,
-  stAssign,     // stitching_assignments
-  asmAssign,    // jeans_assembly_assignments
-  washAssign,   // washing_assignments
-  washInAssign, // washing_in_assignments
-  finAssign     // finishing_assignments
-}) {
-  // placeholders
-  let stitchingStatus="N/A", stitchingOp="", stitchingAssignedOn="N/A", stitchingApprovedOn="N/A";
-  let assemblyStatus= isDenim? "N/A" : "—", assemblyOp="", assemblyAssignedOn="N/A", assemblyApprovedOn="N/A";
-  let washingStatus= isDenim? "N/A" : "—", washingOp="", washingAssignedOn="N/A", washingApprovedOn="N/A";
-  // for non-denim, we skip washing & assembly & washing_in entirely
-  let washingInStatus= isDenim? "N/A": "—", washingInOp="", washingInAssignedOn="N/A", washingInApprovedOn="N/A";
-  let finishingStatus="N/A", finishingOp="", finishingAssignedOn="N/A", finishingApprovedOn="N/A";
-
-  // STITCHING
-  if (!stAssign) {
-    stitchingStatus= "In Cutting";
-    // everything after stitching is "In Cutting"
-    if (isDenim) {
-      assemblyStatus= "In Cutting";
-      washingStatus= "In Cutting";
-      washingInStatus= "In Cutting";
-      finishingStatus= "In Cutting";
-    } else {
-      finishingStatus= "In Cutting";
-    }
-    return {
-      stitchingStatus, stitchingOp, stitchingAssignedOn, stitchingApprovedOn,
-      assemblyStatus, assemblyOp, assemblyAssignedOn, assemblyApprovedOn,
-      washingStatus, washingOp, washingAssignedOn, washingApprovedOn,
-      washingInStatus, washingInOp, washingInAssignedOn, washingInApprovedOn,
-      finishingStatus, finishingOp, finishingAssignedOn, finishingApprovedOn
-    };
-  } else {
-    // we have a stitching assignment
-    const { isApproved, assigned_on, approved_on, opName } = stAssign;
-    stitchingOp= opName|| "";
-    stitchingAssignedOn = assigned_on
-      ? new Date(assigned_on)
-          .toLocaleString('en-GB', { timeZone: 'Asia/Kolkata' })
-          .replace(/\//g, '-')
-      : "N/A";
-    stitchingApprovedOn = approved_on
-      ? new Date(approved_on)
-          .toLocaleString('en-GB', { timeZone: 'Asia/Kolkata' })
-          .replace(/\//g, '-')
-      : "N/A";
-
-    if (isApproved=== null) {
-      stitchingStatus= `Pending Approval by ${stitchingOp}`;
-      if (isDenim) {
-        assemblyStatus= "In Stitching";
-        washingStatus= "In Stitching";
-        washingInStatus= "In Stitching";
-        finishingStatus= "In Stitching";
-      } else {
-        finishingStatus= "In Stitching";
-      }
-      return {
-        stitchingStatus, stitchingOp, stitchingAssignedOn, stitchingApprovedOn,
-        assemblyStatus, assemblyOp, assemblyAssignedOn, assemblyApprovedOn,
-        washingStatus, washingOp, washingAssignedOn, washingApprovedOn,
-        washingInStatus, washingInOp, washingInAssignedOn, washingInApprovedOn,
-        finishingStatus, finishingOp, finishingAssignedOn, finishingApprovedOn
-      };
-    } else if (isApproved==0) {
-      stitchingStatus= `Denied by ${stitchingOp}`;
-      if (isDenim) {
-        assemblyStatus= "In Stitching";
-        washingStatus= "In Stitching";
-        washingInStatus= "In Stitching";
-        finishingStatus= "In Stitching";
-      } else {
-        finishingStatus= "In Stitching";
-      }
-      return {
-        stitchingStatus, stitchingOp, stitchingAssignedOn, stitchingApprovedOn,
-        assemblyStatus, assemblyOp, assemblyAssignedOn, assemblyApprovedOn,
-        washingStatus, washingOp, washingAssignedOn, washingApprovedOn,
-        washingInStatus, washingInOp, washingInAssignedOn, washingInApprovedOn,
-        finishingStatus, finishingOp, finishingAssignedOn, finishingApprovedOn
-      };
-    } else {
-      // approved => check dispatch to next stage
-      if (stitchedQty===0) {
-        stitchingStatus= "In-Line";
-      } else if (isDenim) {
-        // Denim: compare against what assembly received
-        if (!asmAssign) {
-          stitchingStatus= "Completed-Inline";
-        } else if (assembledQty < stitchedQty) {
-          const pend= stitchedQty - assembledQty;
-          stitchingStatus= `${pend} Pending`;
-        } else {
-          stitchingStatus= !washAssign ? "Completed-Inline" : "Completed";
-        }
-      } else {
-        // Hosiery: compare against what finishing received
-        if (!finAssign) {
-          stitchingStatus= "Completed-Inline";
-        } else if (finishedQty < stitchedQty) {
-          const pend= stitchedQty - finishedQty;
-          stitchingStatus= `${pend} Pending`;
-        } else {
-          stitchingStatus= "Completed";
-        }
-      }
-    }
-  }
-
-  // for non-denim, the next step is finishing
-  // for denim: next step is assembly
-  if (!isDenim) {
-    // NON-DENIM => skip assembly, washing, washingIn
-    // finishing next
-    // if there's no finishing assignment, or partial finishing => we do that logic below
-    // keep going...
-  } else {
-    // DENIM => assembly next
-    if (!asmAssign) {
-      assemblyStatus= "In Stitching";
-      washingStatus= "In Stitching";
-      washingInStatus= "In Stitching";
-      finishingStatus= "In Stitching";
-      return {
-        stitchingStatus, stitchingOp, stitchingAssignedOn, stitchingApprovedOn,
-        assemblyStatus, assemblyOp, assemblyAssignedOn, assemblyApprovedOn,
-        washingStatus, washingOp, washingAssignedOn, washingApprovedOn,
-        washingInStatus, washingInOp, washingInAssignedOn, washingInApprovedOn,
-        finishingStatus, finishingOp, finishingAssignedOn, finishingApprovedOn
-      };
-    } else {
-      // we have assembly
-      const { is_approved, assigned_on, approved_on, opName }= asmAssign;
-      assemblyOp= opName|| "";
-      assemblyAssignedOn = assigned_on
-        ? new Date(assigned_on)
-            .toLocaleString('en-GB', { timeZone: 'Asia/Kolkata' })
-            .replace(/\//g, '-')
-        : "N/A";
-      assemblyApprovedOn = approved_on
-        ? new Date(approved_on)
-            .toLocaleString('en-GB', { timeZone: 'Asia/Kolkata' })
-            .replace(/\//g, '-')
-        : "N/A";
-
-      if (is_approved=== null) {
-        assemblyStatus= `Pending Approval by ${assemblyOp}`;
-        washingStatus= "In Assembly";
-        washingInStatus= "In Assembly";
-        finishingStatus= "In Assembly";
-        return {
-          stitchingStatus, stitchingOp, stitchingAssignedOn, stitchingApprovedOn,
-          assemblyStatus, assemblyOp, assemblyAssignedOn, assemblyApprovedOn,
-          washingStatus, washingOp, washingAssignedOn, washingApprovedOn,
-          washingInStatus, washingInOp, washingInAssignedOn, washingInApprovedOn,
-          finishingStatus, finishingOp, finishingAssignedOn, finishingApprovedOn
-        };
-      } else if (is_approved==0) {
-        assemblyStatus= `Denied by ${assemblyOp}`;
-        washingStatus= "In Assembly";
-        washingInStatus= "In Assembly";
-        finishingStatus= "In Assembly";
-        return {
-          stitchingStatus, stitchingOp, stitchingAssignedOn, stitchingApprovedOn,
-          assemblyStatus, assemblyOp, assemblyAssignedOn, assemblyApprovedOn,
-          washingStatus, washingOp, washingAssignedOn, washingApprovedOn,
-          washingInStatus, washingInOp, washingInAssignedOn, washingInApprovedOn,
-          finishingStatus, finishingOp, finishingAssignedOn, finishingApprovedOn
-        };
-      } else {
-        // approved => check dispatch to washing
-        if (assembledQty===0) {
-          assemblyStatus= "In-Line";
-        } else if (!washAssign) {
-          assemblyStatus= "Completed-Inline";
-        } else if (washedQty < assembledQty) {
-          const pend= assembledQty - washedQty;
-          assemblyStatus= `${pend} Pending`;
-        } else {
-          assemblyStatus= !washInAssign ? "Completed-Inline" : "Completed";
-        }
-      }
-
-      // next => washing
-      if (!washAssign) {
-        washingStatus= "In Assembly";
-        washingInStatus= "In Assembly";
-        finishingStatus= "In Assembly";
-        return {
-          stitchingStatus, stitchingOp, stitchingAssignedOn, stitchingApprovedOn,
-          assemblyStatus, assemblyOp, assemblyAssignedOn, assemblyApprovedOn,
-          washingStatus, washingOp, washingAssignedOn, washingApprovedOn,
-          washingInStatus, washingInOp, washingInAssignedOn, washingInApprovedOn,
-          finishingStatus, finishingOp, finishingAssignedOn, finishingApprovedOn
-        };
-      } else {
-        const { is_approved, assigned_on, approved_on, opName }= washAssign;
-        washingOp= opName|| "";
-        washingAssignedOn = assigned_on
-          ? new Date(assigned_on)
-              .toLocaleString('en-GB', { timeZone: 'Asia/Kolkata' })
-              .replace(/\//g, '-')
-          : "N/A";
-        washingApprovedOn = approved_on
-          ? new Date(approved_on)
-              .toLocaleString('en-GB', { timeZone: 'Asia/Kolkata' })
-              .replace(/\//g, '-')
-          : "N/A";
-
-        if (is_approved=== null) {
-          washingStatus= `Pending Approval by ${washingOp}`;
-          washingInStatus= "In Washing";
-          finishingStatus= "In Washing";
-          return {
-            stitchingStatus, stitchingOp, stitchingAssignedOn, stitchingApprovedOn,
-            assemblyStatus, assemblyOp, assemblyAssignedOn, assemblyApprovedOn,
-            washingStatus, washingOp, washingAssignedOn, washingApprovedOn,
-            washingInStatus, washingInOp, washingInAssignedOn, washingInApprovedOn,
-            finishingStatus, finishingOp, finishingAssignedOn, finishingApprovedOn
-          };
-        } else if (is_approved==0) {
-          washingStatus= `Denied by ${washingOp}`;
-          washingInStatus= "In Washing";
-          finishingStatus= "In Washing";
-          return {
-            stitchingStatus, stitchingOp, stitchingAssignedOn, stitchingApprovedOn,
-            assemblyStatus, assemblyOp, assemblyAssignedOn, assemblyApprovedOn,
-            washingStatus, washingOp, washingAssignedOn, washingApprovedOn,
-            washingInStatus, washingInOp, washingInAssignedOn, washingInApprovedOn,
-            finishingStatus, finishingOp, finishingAssignedOn, finishingApprovedOn
-          };
-        } else {
-          // approved => check dispatch to washingIn
-          if (washedQty===0) {
-            washingStatus= "In-Line";
-          } else if (!washInAssign) {
-            washingStatus= "Completed-Inline";
-          } else if (washingInQty < washedQty) {
-            const pend= washedQty - washingInQty;
-            washingStatus= `${pend} Pending`;
-          } else {
-            washingStatus= !finAssign ? "Completed-Inline" : "Completed";
-          }
-        }
-
-        // next => washingIn
-        if (!washInAssign) {
-          washingInStatus= "In Washing";
-          finishingStatus= "In Washing";
-          return {
-            stitchingStatus, stitchingOp, stitchingAssignedOn, stitchingApprovedOn,
-            assemblyStatus, assemblyOp, assemblyAssignedOn, assemblyApprovedOn,
-            washingStatus, washingOp, washingAssignedOn, washingApprovedOn,
-            washingInStatus, washingInOp, washingInAssignedOn, washingInApprovedOn,
-            finishingStatus, finishingOp, finishingAssignedOn, finishingApprovedOn
-          };
-        } else {
-          const { is_approved, assigned_on, approved_on, opName }= washInAssign;
-          washingInOp= opName|| "";
-          washingInAssignedOn = assigned_on
-            ? new Date(assigned_on)
-                .toLocaleString('en-GB', { timeZone: 'Asia/Kolkata' })
-                .replace(/\//g, '-')
-            : "N/A";
-          washingInApprovedOn = approved_on
-            ? new Date(approved_on)
-                .toLocaleString('en-GB', { timeZone: 'Asia/Kolkata' })
-                .replace(/\//g, '-')
-            : "N/A";
-
-          if (is_approved===null) {
-            washingInStatus= `Pending Approval by ${washingInOp}`;
-            finishingStatus= "In WashingIn";
-            return {
-              stitchingStatus, stitchingOp, stitchingAssignedOn, stitchingApprovedOn,
-              assemblyStatus, assemblyOp, assemblyAssignedOn, assemblyApprovedOn,
-              washingStatus, washingOp, washingAssignedOn, washingApprovedOn,
-              washingInStatus, washingInOp, washingInAssignedOn, washingInApprovedOn,
-              finishingStatus, finishingOp, finishingAssignedOn, finishingApprovedOn
-            };
-          } else if (is_approved==0) {
-            washingInStatus= `Denied by ${washingInOp}`;
-            finishingStatus= "In WashingIn";
-            return {
-              stitchingStatus, stitchingOp, stitchingAssignedOn, stitchingApprovedOn,
-              assemblyStatus, assemblyOp, assemblyAssignedOn, assemblyApprovedOn,
-              washingStatus, washingOp, washingAssignedOn, washingApprovedOn,
-              washingInStatus, washingInOp, washingInAssignedOn, washingInApprovedOn,
-              finishingStatus, finishingOp, finishingAssignedOn, finishingApprovedOn
-            };
-          } else {
-            // approved => check dispatch to finishing
-            if (washingInQty===0) {
-              washingInStatus= "In-Line";
-            } else if (!finAssign) {
-              washingInStatus= "Completed-Inline";
-            } else if (finishedQty < washingInQty) {
-              const pend= washingInQty - finishedQty;
-              washingInStatus= `${pend} Pending`;
-            } else {
-              washingInStatus= "Completed";
-            }
-          }
-        }
-      }
-    }
-  }
-
-  // for non-denim, we skip assembly/washing/washingIn entirely
-  // next => finishing
-  if (!finAssign) {
-    if (isDenim) {
-      finishingStatus= "In WashingIn";   // if no finishing assignment
-    } else {
-      finishingStatus= "In Stitching";   // for non-denim
-    }
-    return {
-      stitchingStatus, stitchingOp, stitchingAssignedOn, stitchingApprovedOn,
-      assemblyStatus, assemblyOp, assemblyAssignedOn, assemblyApprovedOn,
-      washingStatus, washingOp, washingAssignedOn, washingApprovedOn,
-      washingInStatus, washingInOp, washingInAssignedOn, washingInApprovedOn,
-      finishingStatus, finishingOp, finishingAssignedOn, finishingApprovedOn
-    };
-  } else {
-    const { is_approved, assigned_on, approved_on, opName }= finAssign;
-    finishingOp= opName|| "";
-    finishingAssignedOn = assigned_on
-      ? new Date(assigned_on)
-          .toLocaleString('en-GB', { timeZone: 'Asia/Kolkata' })
-          .replace(/\//g, '-')
-      : "N/A";
-    finishingApprovedOn = approved_on
-      ? new Date(approved_on)
-          .toLocaleString('en-GB', { timeZone: 'Asia/Kolkata' })
-          .replace(/\//g, '-')
-      : "N/A";
-
-    if (is_approved===null) {
-      finishingStatus= `Pending Approval by ${finishingOp}`;
-    } else if (is_approved==0) {
-      finishingStatus= `Denied by ${finishingOp}`;
-    } else {
-      // partial or complete
-      // for denim => finishing leftover vs washingIn
-      // for non-denim => finishing leftover vs stitched
-      if (isDenim) {
-        if (finishedQty===0) {
-          finishingStatus= "In-Line";
-        } else if (finishedQty>= washingInQty && washingInQty>0) {
-          finishingStatus= "Completed";
-        } else {
-          const pend= washingInQty- finishedQty;
-          finishingStatus= `${pend} Pending`;
-        }
-      } else {
-        // non-denim => finishing leftover vs. stitched
-        if (finishedQty===0) {
-          finishingStatus= "In-Line";
-        } else if (finishedQty>= stitchedQty && stitchedQty>0) {
-          finishingStatus= "Completed";
-        } else {
-          const pend= stitchedQty- finishedQty;
-          finishingStatus= `${pend} Pending`;
-        }
-      }
-    }
-  }
-
-  return {
-    stitchingStatus, stitchingOp, stitchingAssignedOn, stitchingApprovedOn,
-    assemblyStatus, assemblyOp, assemblyAssignedOn, assemblyApprovedOn,
-    washingStatus, washingOp, washingAssignedOn, washingApprovedOn,
-    washingInStatus, washingInOp, washingInAssignedOn, washingInApprovedOn,
-    finishingStatus, finishingOp, finishingAssignedOn, finishingApprovedOn
-  };
-}
-
-/** filterByDept */
-function filterByDept({
-  department, isDenim,
-  stitchingStatus,
-  assemblyStatus,
-  washingStatus,
-  washingInStatus,
-  finishingStatus
-}) {
-  let showRow= true;
-  let actualStatus= "N/A";
-
-  if (department==="all") {
-    if (isDenim) {
-      // finishing if not N/A, else washingIn, else washing, else assembly, else stitching
-      if (!finishingStatus.startsWith("N/A")) actualStatus= finishingStatus;
-      else if (!washingInStatus.startsWith("N/A")) actualStatus= washingInStatus;
-      else if (!washingStatus.startsWith("N/A")) actualStatus= washingStatus;
-      else if (!assemblyStatus.startsWith("N/A")) actualStatus= assemblyStatus;
-      else actualStatus= stitchingStatus;
-    } else {
-      // non-denim => finishing, else stitching
-      if (!finishingStatus.startsWith("N/A")) actualStatus= finishingStatus;
-      else actualStatus= stitchingStatus;
-    }
-    return { showRow, actualStatus };
-  }
-
-  if (department==="cutting") {
-    // always show "Completed"
-    actualStatus= "Completed";
-    return { showRow, actualStatus };
-  }
-
-  if (department==="stitching") {
-    actualStatus= stitchingStatus;
-    return { showRow, actualStatus };
-  }
-
-  if (department==="assembly") {
-    if (!isDenim) return { showRow: false, actualStatus: "N/A" };
-    actualStatus= assemblyStatus;
-    return { showRow, actualStatus };
-  }
-
-  if (department==="washing") {
-    if (!isDenim) return { showRow: false, actualStatus: "N/A" };
-    actualStatus= washingStatus;
-    return { showRow, actualStatus };
-  }
-
-  if (department==="washing_in") {
-    if (!isDenim) return { showRow: false, actualStatus: "N/A" };
-    actualStatus= washingInStatus;
-    return { showRow, actualStatus };
-  }
-
-  if (department==="finishing") {
-    actualStatus= finishingStatus;
-    return { showRow, actualStatus };
-  }
-
-  return { showRow, actualStatus };
-}
 
 /** The final PIC Report route */
 /*******************************************************************
@@ -2892,11 +1963,6 @@ async function sendChunk() {
 }
 
 /** Returns how many days since the dateValue. */
-function daysSince(dateValue) {
-  if (!dateValue) return 0;
-  const msDiff = Date.now() - new Date(dateValue).getTime();
-  return Math.floor(msDiff / (1000 * 60 * 60 * 24));
-}
 
 /**
  * GET  /urgent-tat   => Show a page with previews + single "Send" button
